app/requirements.txt
numpy
pandas
matplotlib
seaborn
lime
pillow
joblib
requests
